Don't try to reclaim deallocated blocks on a hashcache hit
This behavior would be nice, but it was causing persistent hangs.  Specifically if you write a block,
then delete the file, then immediately write the same data to the same block, you get a cache hit.
Something in there was not working right at that point, thus when you try to rm that file it hangs. This was
repeatable.  Instead inside try_dedup_block only attempt to do the dedup if the found block is different
from the allocated block and has a ref count > 0.